+++ Tests/RetentionFixtures/testRetainImplicitDeclarations.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct FixtureStruct2 {
+    let someVar: String
+}